+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - The new statement for the checking logical assertions.
 
+## [0.12.0] - 2022-02-13
+
+### Removed
+
+- The `IncludeVarsLocal` annotation.
+- The `TargetAnnotationTypes` annotation.
+
 ## [0.11.1] - 2022-01-31
 
 ### Fixed